Bullet Meteor hit complete
The bullet succesfully destroys itself and the meteor divides into 2
smaller meteors with a bit smaller speed and a split in direction.

Next step is to create a better collision, this is probaly done by
adjusting the base x,y pos to the center of the object. Maybe giving the
element a negative left, and top margin will be an easy solution for
this.