General adjustments
Adjusted the x & y pos to center the object. And fixed the problems with
wrapping caused by this adjustment.
Also did some small adjustments like setting a max # bullets and started
with making the Asteroids looking nice when the size in- or decreases.

Next: Work further with the looks of Astereroids when sizing in- or
decreases and try to improve the keys (moving+shooting doesn't work
smooth)

Next Step: